+++ glesgae/gae_memory.c
@@ -2,12 +2,12 @@
 
 #include <stdlib.h>
 
-void* gae_alloc(unsigned int size)
+void* gae_alloc(unsigned long size)
 {
 	return malloc((size_t)size);
 }
 
-void* gae_realloc(void* ptr, unsigned int size)
+void* gae_realloc(void* ptr, unsigned long size)
 {
 	return realloc(ptr, (size_t)size);
 }
