+++ glesgae/gae_memory.h
@@ -2,10 +2,10 @@
 #define _GAE_MEMORY_H_
 
 /* Allocate a new chunk of memory */
-void* gae_alloc(unsigned int size);
+void* gae_alloc(unsigned long size);
 
 /* Reallocate a chunk of memory */
-void* gae_realloc(void* ptr, unsigned int size);
+void* gae_realloc(void* ptr, unsigned long size);
 
 /* Free a chunk of memory */
 void gae_free(void* ptr);
